ajout des trois principale actions de FIFO  get, add, new
